+++ R/add_tbv.R
@@ -104,13 +104,23 @@ add_tbv <- function(tbl, trait, date_calc = Sys.Date()) {
     a <- genome[[paste0("add_", t)]]
     a[is.na(a)] <- 0
 
+    p_base_col <- paste0("base_allele_freq_", t)
+    p_base <- if (p_base_col %in% names(genome)) {
+      pv <- as.numeric(genome[[p_base_col]])
+      pv[is.na(pv)] <- 0
+      pv
+    } else {
+      rep(0, length(a))
+    }
+
     if (m$expressed_parent == "both") {
       rows_idx <- match(ids_t, rownames(geno_mat_full))
-      tbv <- as.numeric(geno_mat_full[rows_idx, , drop = FALSE] %*% a)
+      tbv <- as.numeric(geno_mat_full[rows_idx, , drop = FALSE] %*% a) -
+             2 * sum(p_base * a)
     } else {
       parent_origin <- if (m$expressed_parent == "parent_1") 1L else 2L
       hap_mat <- get_haplotype_matrix(pop, parent_origin, ids_t)
-      tbv <- as.numeric(hap_mat %*% a)
+      tbv <- as.numeric(hap_mat %*% a) - sum(p_base * a)
     }
 
     tbv_df <- tibble::tibble(